+++ packages/terra-form-select/src/shared/_Option.jsx
@@ -3,9 +3,6 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import classNamesBind from 'classnames/bind';
 import ThemeContext from 'terra-theme-context';
-import { injectIntl } from 'react-intl';
-import { IconAdd } from 'terra-icon';
-import Button from 'terra-button';
 import SharedUtil from './_SharedUtil';
 import styles from './_Option.module.scss';
 
@@ -69,11 +66,6 @@ const propTypes = {
    * The i18n value of the text "Expanded combobox".
    */
   expandedStateText: PropTypes.string,
-  /**
-   * @private
-   * The intl object containing translations. This is retrieved from the context automatically by injectIntl.
-   */
-  intl: PropTypes.shape({ formatMessage: PropTypes.func }).isRequired,
 };
 
 const defaultProps = {
@@ -94,7 +86,6 @@ const Option = ({
   totalOptions,
   ofText,
   expandedStateText,
-  intl,
   ...customProps
 }) => {
   const theme = React.useContext(ThemeContext);
@@ -140,8 +131,7 @@ const Option = ({
       data-terra-select-option
       aria-label={itemLabel}
     >
-      {isAddOption && <Button className={cx('add-button')} text={intl.formatMessage({ id: 'Terra.form.select.addText' })} icon={<IconAdd />} />}
-      {(isCheckable && !isAddOption) && <span className={cx('icon')} />}
+      {(isCheckable || isAddOption) && <span className={cx('icon')} />}
       <span data-terra-add-option className={cx('display')}>{display}</span>
     </li>
   );
@@ -151,4 +141,4 @@ Option.propTypes = propTypes;
 Option.defaultProps = defaultProps;
 Option.isOption = true;
 
-export default injectIntl(Option);
+export default Option;