+++ packages/terra-tag/src/index.js
@@ -1,0 +1,5 @@
+import TagList from './TagList';
+import Tag from './Tag';
+
+export default Tag;
+export { TagList };