+++ packages/terra-form-select/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* Changed
+  * Change role from application to combobox in `terra-form-select-combobox`.
+  * Replace add text to button in dropdown list in  `terra-form-select-combobox`.
+  * Ability to close dropdown and clear text by ESC key in `terra-form-select-combobox`.
+
 * Added
   * Added No results found text in `terra-form-select-combobox`.