+++ packages/terra-form-select/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Added
+  * Added results that contain in `terra-form-select-combobox`.
   * Added No results found text in `terra-form-select-combobox`.
 
 ## 6.49.0 - (October 23, 2023)