+++ packages/terra-form-select/CHANGELOG.md
@@ -7,6 +7,7 @@
   * Ability to close dropdown and clear text by ESC key in `terra-form-select-combobox`.
 
 * Added
+  * Added results that contain in `terra-form-select-combobox`.
   * Added No results found text in `terra-form-select-combobox`.
 
 ## 6.49.0 - (October 23, 2023)